// portfolio-frontend/src/app/admin/models/dashboard-stats.ts
export interface DashboardStats {
    totalProjects: number;
    totalBlogs: number;
    totalProjectViews: number;
    totalProjectLikes: number;
    publishedBlogs: number;
    featuredProjects: number;
    totalGithubClicks: number;
    totalDemoClicks: number;
    totalDetailClicks: number;
    totalComments: number;
    totalVisitors: number;
}